refactor/ gap을 구하는 유틸 이름 가독성 좋게 수정
