+++ src/components/SelectionLayer.vue
@@ -11,8 +11,8 @@ import { useSelectionStore } from '@/store/selection';
 import Line from '@/classes/Line';
 import getLineNum from '@/utils/getLineNum';
 import hasText from '@/utils/hasText';
-import getFirstLineStartGap from '@/utils/line/getFirstLineStartGap';
-import getLastLineEndGap from '@/utils/line/getLastLineEndGap';
+import getLeftGap from '@/utils/line/getLeftGap';
+import getRightGap from '@/utils/line/getRightGap';
 import getSelectedTokens from '@/utils/getSelectedTokens';
 import getLineMap from '@/utils/line/getLineMap';
 
@@ -79,8 +79,8 @@ function drawSelection(range: Range) {
 
     if (!startLine || !endLine) return;
 
-    const startGap = getFirstLineStartGap(startContainer, startOffset);
-    const endGap = getLastLineEndGap(endContainer, endOffset);
+    const startGap = getLeftGap(startContainer, startOffset);
+    const endGap = getRightGap(endContainer, endOffset);
     startLine.setStartPos({ x: startLine.left + startGap, y: startLine.top });
     endLine.setEndPos({ x: endLine.right - endGap, y: endLine.bottom });
