--- src/utils/line/getFirstLineStartGap.ts
@@ -1,11 +1,0 @@
-export default function getFirstLineStartGap(
-    startContainer: Node,
-    startOffset: number
-) {
-    const beforeStartRange = new Range();
-    beforeStartRange.setStart(startContainer, 0);
-    beforeStartRange.setEnd(startContainer, startOffset);
-    const { width } = beforeStartRange.getBoundingClientRect();
-
-    return width;
-}
