--- src/utils/line/getLastLineEndGap.ts
@@ -1,14 +1,0 @@
-export default function getLastLineEndGap(
-    endContainer: Node,
-    endOffset: number
-) {
-    const afterEndRange = new Range();
-    const wholeText = (endContainer as Text).wholeText;
-
-    afterEndRange.setStart(endContainer, endOffset);
-    afterEndRange.setEnd(endContainer, wholeText.length);
-
-    const { width } = afterEndRange.getBoundingClientRect();
-
-    return width;
-}
